+++ composer.json
@@ -21,6 +21,7 @@
     "slim/slim": "4.*",
     "symfony/console": "^6.3",
     "symfony/error-handler": "^6.3",
+    "symfony/filesystem": "^6.3",
     "symfony/process": "^6.3",
     "symfony/uid": "^6.3",
     "vlucas/phpdotenv": "^5.5"
