--- src/Filesystem/LocalSystemException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Takemo101\Chubby\Filesystem;
-
-use Exception;
-
-/**
- * Exception regarding LocalSystem.
- */
-class LocalSystemException extends Exception
-{
-    //
-}
