+++ tests/Filesystem/FilesystemTestCase.php
@@ -4,6 +4,8 @@ namespace Tests\Filesystem;
 
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use Takemo101\Chubby\Filesystem\LocalFilesystem;
+use Takemo101\Chubby\Filesystem\PathHelper;
+use Takemo101\Chubby\Filesystem\SymfonyLocalFilesystem;
 
 class FilesystemTestCase extends BaseTestCase
 {
@@ -26,7 +28,7 @@ class FilesystemTestCase extends BaseTestCase
      */
     public function setUpLocalFilesystem(): void
     {
-        $this->filesystem = new LocalFilesystem();
+        $this->filesystem = new SymfonyLocalFilesystem();
     }
 
     /**
@@ -37,7 +39,7 @@ class FilesystemTestCase extends BaseTestCase
      */
     public function getTestResourcePath(string $file = ''): string
     {
-        $helper = $this->filesystem->helper;
+        $helper = new PathHelper();
 
         $path = $helper->join(dirname(__DIR__, 1), 'resource/filesystem');
